+++ default_user_info_service.go
@@ -3,11 +3,11 @@ package auth
 import "context"
 
 type DefaultUserInfoService struct {
-	MaxPasswordAge int64
+	MaxPasswordAge int32
 	*BaseUserInfoService
 }
 
-func NewUserInfoService(authenticationRepository AuthenticationRepository, maxPasswordAge int64, maxPasswordFailed int, lockedMinutes int) *DefaultUserInfoService {
+func NewUserInfoService(authenticationRepository AuthenticationRepository, maxPasswordAge int32, maxPasswordFailed int, lockedMinutes int) *DefaultUserInfoService {
 	b := NewBaseUserInfoService(authenticationRepository, maxPasswordFailed, lockedMinutes)
 	return &DefaultUserInfoService{maxPasswordAge, b}
 }
